Make reloads non-blocking
Consider a key `k` that is already loaded into cache, and suppose we
call `Cache.request_load(k)`.

The existing PDL behavior is to place the key into `:requested` state,
which will block any subsequen `get` requests for that key until it is
loaded.

The behaviour introduced in this PR is to place the key into a `:reload`
state; `get` requests will be non-blocking (returning the already-cached
value) during the reload.  The load worker now prioritizes `:requested`
keys over `:reload` keys.